+++ src/components/Inputs/ColorPicker/types.tsx
@@ -15,11 +15,11 @@ export type ColorPickerChangeEvent = MInputChangeEvent<ColorValue>;
 
 export type ColorPickerProps<TReturnedColor extends ReturnedColor> = InputProps & {
   /** Color value for the picker.
-   * - `ColorValue | null` those are the only possible types for controlled value.
-   * - If `undefined` will be passed as value, component will be in uncontrolled mode.
+   * - If `ColorValue` is passed as value, component will be in `controlled` mode.
+   * - If `undefined` is passed as value, component will be in `uncontrolled` mode.
    * @default undefined
    */
-  value?: ColorValue | null;
+  value?: ColorValue;
 
   /** Default color value for the picker. */
   defaultValue?: ColorValue;
@@ -40,7 +40,7 @@ export type ColorPickerProps<TReturnedColor extends ReturnedColor> = InputProps 
   onChange?: (event: ColorPickerChangeEvent, value: ColorValue) => void;
 
   /** Callback triggered when the color picker closes. */
-  onClose?: (value: ColorValue | null) => void;
+  onClose?: (value: ColorValue) => void;
 };
 
 export type HslValue = {