+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - [[#186]](https://github.com/Marcin-Migdal/m-component-library/issues/186) **[PATCH]** Fix multiple bug 186, 187, 191, export FormErrorsType, Form register and registerBlur type improvement, changeDropdown input optionMenu z-index
 - [[#188]](https://github.com/Marcin-Migdal/m-component-library/issues/188) **[PATCH]** Fix overlay disabled key press, add onClose callback that is called on "esc" button press.
+- [[#189]](https://github.com/Marcin-Migdal/m-component-library/issues/189) **[PATCH]** Controlled value of the ColorPicker cannot be null, now for controlled input value has to be clearly and properly defined.
 
 ### Release date 20.03.2025 - Release 2.1.1